add receiptRoute, transactions & uploadReceipt API

+++ migrations/20221115143337_create-transcations.ts
@@ -1,6 +1,5 @@
 import { Knex } from "knex";
 
-
 export async function up(knex: Knex): Promise<void> {
     await knex.schema.createTable("transcations", (table) => {
         table.increments();
@@ -8,14 +7,12 @@ export async function up(knex: Knex): Promise<void> {
         table.foreign("debitor_id").references("users.id");
         table.integer("creditor_id").notNullable();
         table.foreign("creditor_id").references("users.id");
-        table.integer("transcations_amount").notNullable();
+        table.float("transcations_amount").notNullable();
         table.boolean("is_settled");
         table.timestamps(false, true);
-      });
+    });
 }
-
 
 export async function down(knex: Knex): Promise<void> {
     await knex.schema.dropTableIfExists("transcations");
 }
-
